+++ test/inet/data/ip_test.clj
@@ -1,8 +1,12 @@
 (ns inet.data.ip-test
-  (:require [inet.data.ip :as ip])
+  (:require [clojure.edn :as edn]
+            [clojure.java.io :as io]
+            [inet.data.ip :as ip])
   (:use [clojure.test])
   (:import [java.net InetAddress]
            [inet.data.ip IPException IPNetworkException]))
+
+(set! *warn-on-reflection* true)
 
 (deftest test-strict-network-operations-reject-invalid-input
   (is (thrown? IPException
@@ -215,9 +219,10 @@
 (deftest test-network-set-to-array-jdk-11
   (let [expected [(ip/network "10.0.0.0/24")
                  (ip/network "10.0.1.0/24")]
-        networks (ip/network-set "10.0.0.0/24" "10.0.1.0/24")]
-    (is (= expected (vec (.toArray networks))))
-    (is (= expected (vec (.toArray networks (into-array expected)))))
+        networks (ip/network-set "10.0.0.0/24" "10.0.1.0/24")
+        ^objects target (into-array Object expected)]
+    (is (= expected (vec (.toArray ^java.util.Collection networks))))
+    (is (= expected (vec (.toArray ^java.util.Collection networks target))))
     (is (= expected (vec networks)))
     (is (= expected (into [] networks)))))
 
@@ -268,6 +273,51 @@
 (defn covered-addresses
   [nets]
   (set (mapcat seq nets)))
+
+(defn network-subtract
+  [a b]
+  (if-let [subtract (ns-resolve 'inet.data.ip 'network-subtract)]
+    (subtract a b)
+    ::missing))
+
+(defn network-intersect
+  [a b]
+  (if-let [intersect (ns-resolve 'inet.data.ip 'network-intersect)]
+    (intersect a b)
+    ::missing))
+
+(deftest test-network-intersect
+  (testing "returns the narrower network for overlapping CIDRs"
+    (is (= (ip/network "10.1.0.0/16")
+           (network-intersect "10.0.0.0/8" "10.1.0.0/16")))
+    (is (= (ip/network "2001:db8::/128")
+           (network-intersect "2001:db8::/126" "2001:db8::/128"))))
+  (testing "returns the identical network for equality"
+    (is (= (ip/network "10.0.0.0/24")
+           (network-intersect "10.0.0.0/24" "10.0.0.0/24"))))
+  (testing "returns nil for disjoint and mixed-family networks"
+    (is (nil? (network-intersect "10.0.0.0/24" "10.0.1.0/24")))
+    (is (nil? (network-intersect "10.0.0.0/24" "2001:db8::/64")))))
+
+(deftest test-network-subtract
+  (testing "subtracts a contained IPv4 network into minimal complement blocks"
+    (is (= (ip/network-set "10.0.0.0/16" "10.2.0.0/15" "10.4.0.0/14"
+                           "10.8.0.0/13" "10.16.0.0/12" "10.32.0.0/11"
+                           "10.64.0.0/10" "10.128.0.0/9")
+           (network-subtract "10.0.0.0/8" "10.1.0.0/16"))))
+  (testing "works for IPv6"
+    (is (= (ip/network-set "2001:db8::/127")
+           (network-subtract "2001:db8::/126" "2001:db8::2/127"))))
+  (testing "handles disjoint, containing, and equal networks"
+    (is (= (ip/network-set "10.0.0.0/24")
+           (network-subtract "10.0.0.0/24" "10.0.1.0/24")))
+    (is (= (ip/network-set)
+           (network-subtract "10.0.0.0/24" "10.0.0.0/16")))
+    (is (= (ip/network-set)
+           (network-subtract "10.0.0.0/24" "10.0.0.0/24"))))
+  (testing "treats mixed-family networks as disjoint"
+    (is (= (ip/network-set "10.0.0.0/24")
+           (network-subtract "10.0.0.0/24" "2001:db8::/64")))))
 
 (deftest test-aggregate-networks
   (testing "absorbs networks contained by another"
@@ -348,6 +398,19 @@
         (is (= expected (ip/special-use representative)) representative)
         (is (= expected (ip/special-use first)) first)
         (is (= expected (ip/special-use last)) last)))))
+
+(deftest test-special-use-registry-resource
+  (let [resource (io/resource "inet/data/special-use-registry.edn")]
+    (is resource "special-use registry is on the classpath")
+    (when resource
+      (let [registry (edn/read (java.io.PushbackReader. (io/reader resource)))]
+        (is (= "repository-snapshot-2026-08-23" (:registry-version registry)))
+        (is (= ["IPv4 Special-Purpose Address Space"
+                "IPv6 Special-Purpose Address Space"]
+               (:source-registries registry)))
+        (is (some #(= {:name :private :network "10.0.0.0/8"} %) (:blocks registry)))
+        (is (= :private (ip/special-use "10.1.2.3")))
+        (is (= :documentation (ip/special-use "2001:db8::1")))))))
 
 (deftest test-special-use-predicates
   (testing "Category predicates"